requirements.txt
click>=8.1.*
pytest>=8.3.*
pytest-mock>=3.14.*
PyYAML==6.0.*
psycopg2-binary==2.9.*
requests>=2.32.*